Make resolve into a macro
