+++ myia/operations.py
@@ -57,6 +57,7 @@ make_tuple = MyiaOnlyOperation('make_tuple')
 make_list = MyiaOnlyOperation('make_list')
 make_dict = MyiaOnlyOperation('make_dict')
 next = MyiaOnlyOperation('next')
+resolve = MyiaOnlyOperation('resolve')
 slice = MyiaOnlyOperation('slice')
 switch = MyiaOnlyOperation('switch')
 to_array = MyiaOnlyOperation('to_array')
