+++ myia/pipeline/resources.py
@@ -52,6 +52,7 @@ scalar_object_map = {
     operations.iter: C.iter,
     operations.hasnext: C.hasnext,
     operations.next: C.next,
+    operations.resolve: M.resolve,
     operations.to_array: C.to_array,
     operations.switch: P.switch,
     operations.user_switch: M.user_switch,
@@ -108,6 +109,7 @@ standard_object_map = {
     operations.iter: C.iter,
     operations.hasnext: C.hasnext,
     operations.next: C.next,
+    operations.resolve: M.resolve,
     operations.to_array: C.to_array,
     operations.switch: P.switch,
     operations.user_switch: M.user_switch,
